requirements.txt
opencv-python
numpy
pillow
streamlit
pyserial
matplotlib